requirements-fitdev.txt
# Core dependencies
crewai>=0.30.0
crewai-tools>=0.2.0
python-dotenv>=1.0.0

# Language models
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.2

# Web capabilities
requests>=2.31.0
beautifulsoup4>=4.12.0

# Data handling
pydantic>=2.0.0

# Testing
pytest>=7.0.0

# Development tools
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.0.0
